+++ pages/api/auth/[...nextauth].ts
@@ -80,6 +80,9 @@ export const authOptions: NextAuthOptions = {
 
       return session
     }
+  },
+  pages: {
+    signIn: '/login'
   }
 }